requirements.txt
# Environment Variable Management
python-dotenv

# web dev
streamlit

# LangChain 
langchain
langchain-core
langchain-community
langchain-chroma
langchain-google-genai

# vector database
chromadb

# api
fastapi
uvicorn
requests
pydantic 